backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-telegram-bot==21.10
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet>=3.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
